README>MD
Code for home-made ESP8266 Air conditioner controller. This repository has an android app that sends JSON messages via MQTT to the controllers. 
It also has a folder for the arduino code for the ESP8266 controllers. 
Notes on the project at https://hackaday.io/project/13279-esp8266-mqtt-infrared-aircon-control 
It is a work in progress, and at present the code is not cleaned up.